refactor: move out paths config
