+++ configs/config_api.py
@@ -5,7 +5,7 @@ It will be handle by configs/main.py
 """
 
 from pathlib import Path
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
 
 @dataclass(frozen=True)
